feat: class with the methods to create the table is done
